requirements.txt
opencv-python
matplotlib
pyyaml
tqdm
requests
psutil
dill
pywavelets
timm
scipy
thop 
einops
pandas
seaborn
ultralytics-thop
polars